Add PKGBUILD for orbisgis daily snapshot
The problem with this is that the md5 sums will expire every day,
and it would be too much work to update them all the time.

A better technique would be to do a git checkout of master. But then
we would have to build from source. Add maven to makedepends?